Source code for the code-reuse agent (so I don't lose it).
This is a work-in-progress.  DO NOT USE.  (I guarantee that it does not work right now.)

Much of this directory are temporary files and work notes.
These will be deleted.
And the main files will be renamed to something more useful
than "fc.c" for "filter C and C++ code".

The current game plan is to rewrite all of the perl code
into C and interface it into the database.

Questions? Ask Neal Krawetz.



git-svn-id: http://svn.code.sf.net/p/fossology/code/trunk@353 a5643f74-2041-0410-a2b9-94ffef304a07