Fixed the preg_replace pattern to be less specific.  Also investigated and put in a comment a more general expression that should not fail. Sirius used System Repository for it's root folder name and the default install name is Software Repository.  It could be anything, hence the suggestion to use the most general pattern.
This version still has a defect.  If you use 'Software Repository' (or
the root folder name), it will create a subfolder by the same name and
create a circular folder loop in the foldercontents table.  This then
breaks browsing (runs out of process memory) and all uploads from the
ui, which need the browse functions to work.  This appears to be a
flaw in the way that GetFolder works.

Another defect is that duplicate upload folder names can be created in
the same folder.

Need to talk with Neal and Bob on the above two and they are on FTO.


git-svn-id: http://svn.code.sf.net/p/fossology/code/trunk@1041 a5643f74-2041-0410-a2b9-94ffef304a07