Fixed issues with the sentence boundary model. Due to an error events were only getting recorded once. This meant that if you had an event with opposing labels ('.\n\n' labeled as both a sentence break and not) the probabilities were recorded as 0.5 and 0.5 even if the true probabilities were 0.99 and 0.01.
test.cpp was moved to test_app.cpp in order to create a set of unit tests.

repr.* was added to escape text when printing and creating features for the maxent model.

hash.h holds two hash functions. One for char* and the other for c++ strings.

f1.cpp still prints the wrong license index when reporting to f1.php. This results in seemingly incorrect labeling of license sections. This should be fixed soon.


git-svn-id: http://svn.code.sf.net/p/fossology/code/trunk/project@2526 a5643f74-2041-0410-a2b9-94ffef304a07