New tokenizing code. New model tuning code.
All possible tokens are now located using the same regexp. This fixes a major problem when analyzing binary files where the regexp get stuck on a large section, and the scheduler would then restart the agent. The new tokenizor also improves the general speed of the agent. Since only one regexp is used an instant improvement of 6X can be seen.

Added a new library function for tuning the naive bayes model. This does a sequence of leave one out evaluations of the training data to converge on approximately optimal values of the prior probabilities and the Laplacian smoothing parameters.


git-svn-id: http://svn.code.sf.net/p/fossology/code/trunk/fossology@3128 a5643f74-2041-0410-a2b9-94ffef304a07