Changed ui-browse and search-file to use sysconfig variables.  These changes enable not displaying search and browse when no one is logged in.  Added the variables to admin-config.php.
There is a dependency in search-file on browse.  So when public browse
is turned off, search gets turned off too.

If browse is turned off, and search is turned on, search will appear
AFTER the user logs in.

If both browse and search is ON (true), then they appear when no one
is logged in.

If both browse and search are off (false), then neither appear when
no-one is logged in, but browse appears afterlogging in.  Search never
appears except in the micro menu.

Everything else works the same.

git-svn-id: http://svn.code.sf.net/p/fossology/code/trunk/fossology@4009 a5643f74-2041-0410-a2b9-94ffef304a07