Change all the logging functions by adding a LOG_ prefix. This is because defines like "DEBUG" are common in code and will cause a name collision.  Adding LOG_ also makes it a little clearer what the purpose of the macro is.
Add LOG_NOTICE

Add LOG_ to all the verbose macros.

Update the verbose macro comments so it is clear what they are for.

Change the name of the verbose macros so that the trailing digit is the 
power of 2 of the bit mask.

Add additional verbosity levels, there are now 8.

Change the global "verbose" to "agent_verbose" to be more specific and avoid
other name collisions with our current code.



git-svn-id: http://svn.code.sf.net/p/fossology/code/branches/fossology2.0@4833 a5643f74-2041-0410-a2b9-94ffef304a07