Changes for fo-postinstall and friends.
fo-postinstall: added {$LIBEXECDIR} to invocations of fossinit and
fo_dbcheck.  Changed to call fossinit.php instead of fossinit.
Changed chgrp of Db.conf so it now works.  To get this to work added
PROJECTGROUP to fossology.conf.in.

Default user and fossy were not getting created in the db on an
install.  Issue was that code was removed from libschema that was
calling the install methods of every plugin.  This will not work in
fossinit as the plugins have not yet been loaded.  Core-auth's Install
method is where the default user and fossy get created.

Added function plugin_install to common-plugin, this is used in
index.php to call the install methods of all the plugins.  Modified index.php.

Removed adjustments for simple ui from core-auth.php, it was trying to
use simpleindex.php which is no longer available.

Added fo_dbcheck.php to svn.

Added installation of fo_dbcheck and fossinit to LIBEXECDIR in the
install/Makefile.  These modules should be there according to the
standard and where fo-postinstall expects them to be located.

Bob added -c paramater and it's processing to fossinit and refactored
some code.


git-svn-id: http://svn.code.sf.net/p/fossology/code/branches/fossology2.0@5150 a5643f74-2041-0410-a2b9-94ffef304a07