Bug fix http://www.fossology.org/issues/615
Windows executables look like archives and 7z will attempt to unpack them.
If that happens there will be a .bss file representing the .bss segment.
7z will try to unpack this further, potentially getting into an infinite
unpack loop.  So if you see an octet type .bss file, consider it text
to avoid this problem.  This problem was first noticed on papi-4.1.3-3.el6.src.rpm


git-svn-id: http://svn.code.sf.net/p/fossology/code/branches/fossology2.0/fossology@5448 a5643f74-2041-0410-a2b9-94ffef304a07