feat(Obligation): add first implementation of obligations and risks management
Create two new tables: "obligation_ref" and "obligation_map".
The table "obligation_ref" contains the topic and text of the obligations.
The table "obligation_map" maps obligation references to licenses.

The combination of an obligation topic (ob_topic) and text (ob_text)
must be unique. One obligation can be mapped to several licenses.
The license shortnames are used in the UI, only licenses already
available in the "license_ref" table can be used. The UI uses javascript
code for the license selection, the implementation is based on "select2".

CSV export and import functions are available. The license list
is exported as a string with ";" as separator.