fix(browse): ajax browse required login
The browse plugin `ui-browse.php` does not require login to be used, but the
ajax backend `AjaxBrowse.php` required it.

The problem was caused due to the migration of the parent class `FO_Plugin` to
`DefaultPlugin` in the `AjaxBrowse.php`. In `FO_Plugin` is by default no login
required, this logic is flipped in the newer implementation `DefaultPlugin`.

closes #994