Fix(ununpack): Fix compiler warnings for Debian 10/gcc8
It seems Debian 10/gcc8 8.3.0-6 is even more strict than
Ubuntu Bionic gcc 8.3.0-6.

In case of strncpy we should know what we are doing, but gcc can still
throw warnings because usage of truncating strncpy is very rarely 'correct'
and the warning is also about the possibly missing 0 termination.
Better way to suppress the warning is to use memcpy and add the 0 termination.

Signed-off-by: Andreas J. Reichel <andreas.reichel@tngtech.com>