feat(notice_import): Import notice file content into acknowledgement
This feature enables users to add the contents of NOTICE* files within
the uploaded repo tree into the Acknowledgement field for license clearing.
To accomplish this, a new modal dialog was designed.

Extra: The user can add the text fragments of more than one NOTICE file at
the current cursor position of the TextArea input.

Signed-off-by: Andreas J. Reichel <andreas.reichel@tngtech.com>