Changes to install section of README
I've made some changes to the installation section of the README.md file in an effort to make it easier and clearer. I think at the moment, it is not as clear to install via containers and it might be. One of the reasons is that if one chooses to install a single container, you get a scary warning that data is not persistent. It looks like there's a lot to do in that single container setup to make data persist, so my recommendation would be to use the Docker compose method as the sole container method for installation. I've added the -d switch to the description of the docker compose method and I've separated it into separate instructions because the previous instructions ran together. Happy to make changes as needed and to discuss.