fix(delete): Do not remove upload_pk
While deleting a upload, due to constraint `upload_pfile_fk_fkey` the
tupple from upload table also get dropped. This causes issues with the
job table as they hold the upload_fk which causes issue for new uploads.

This commit drops that constraint, retaining the upload_pk and marks the
upload as deleted using the `expire_action` and `expire_date` columns.
Since the upload link is removed from `foldercontents` table, no other
changes were required. The commit also adds a filter to removeUploads()
option for maintenance agent so that it does not drop the tupple.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>