fix(upload): Fix the check for expire_action
The PR #1707 included changes which changed the behaviour of
expire_action column of upload table. A new check was added to check if
the upload was deleted. The check looks for 'd' in expire_action column
of upload table. The default value for the column in NULL which fails
the check of `!= 'd'` in Postgres.

This commits fixes the check by additional check of column being NULL.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>