feat(rest): Filter uploads by folder id
This PR allows user to limit uploads results to given folder id, provide
an option to list uploads recursively, paginate the results if more than
100 uploads exists (can be changed with header parameter).

The default folder id is the user's root folder id and recursive is set
to true.

In the earlier implementation, the uploads listing was limited to the
user accessing the endpoint meaning that you can list only the uploads
done by you. However, the UI lists the uploads based on group id. The
REST API now reflects the same behaviour.

The listing of `groupName` parameter is updated in the `openapi.yaml` to
reduce the redundancy and add the endpoints which were missing it.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>