feat(build): Get build version number from git
Use the information provided by git describe --tags and generate the
build version number.

The `git describe --tags` also displays the number of commits since last
tags which is used here to get the build number. So let's say, last tag
was 3.8.1 and there are 88 commits since then, the version now will be
3.8.88.
(See `man git-describe` for info on format)

Update the fo-mktar to use this version info and a new script fo-debuild
to generate debian packages and create an archive from them.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>