feat(password): Create password policy
Allow admins to enforce password policies from Customize page. Following
options can be chosen to generate policy:
1. Enable/disable policy check.
2. Fixing length of passwords (min and max).
3. Requiring minimum one of lowercase character.
4. Requiring minimum one of uppercase character.
5. Requiring minimum one digit.
6. Requiring minimum one of special characters from a predefined set.

This policy is used to generate RegEx and the password will be checked
while creating a new user or editing a user.

The checks are added to the backend as well as frontend.

Also, if password policy is enabled, do not allow empty passwords for
new users.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>