fix(cd): Use published event to build release pkgs
The created event type works only if a release was done without making a
draft. But published type event is raised when a release, pre-release or
a draft of a release is published. This will make sure the pipeline runs
irrespective of draft state.

For more, read: https://docs.github.com/en/free-pro-team@latest/developers/webhooks-and-events/webhook-events-and-payloads#release

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>