fix(decision): Create ce for folder decisions irr
When someone marked a whole folder as irrelevant, the clearing_events
for the files went missing. This causes data inconsistency while getting
the license for those clearing_decisions.

This PR uses ClearingDecisionProcessor::makeDecisionFromLastEvents() to
generate clearing_decision and clearing_event.
Consequently, the functions
markDirectoryAsDecisionTypeIfScannerDetected() and
markDirectoryAsDecisionTypeIfUserEdited() were no longer required to be
separate and thus replaced with markDirectoryAsDecisionTypeRec().

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>