feat(dbcreate): retry psql check while starting
Sometimes when we stop and start the docker instance.
We get the following error message:
psql: FATAL:  the database system is starting up
ERROR: postgresql isn't running

It doesn't mean that the postgresql is not running.
It means postgresql is preparing to start.
So we might need to check for a while and wait.

Signed-off-by: Ying-Chun Liu (PaulLiu) <paulliu@debian.org>