feat(monkbulk): Configurable irrelevant file scan
Allow user to choose if they want to ignore irrelevant files while doing
a monkbulk scan or not.

The existing scanner ignores irrelevant files by default but works on
pfile. While inserting the data back, this pfile needs to be translated
into uploadtree_pk. During this translation, if a pfile occurs multiple
times and is marked as irrelevant in some places but not in other, can
create multiple uploadtree_pk which is undesirable.
This PR fixes that by checking clearing_decision while inserting new
results as well.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>