feat(rest): Update upload information
Use PATCH to update the update the upload information.

New parameters `assignee` and `status` can be passed as query to the
/uploads/{id} endpoint as PATCH request.
The comment for Rejected and Closed status can be sent as body of
`text/plain` type.

BREAKING CHANGE: The PATCH was used for copy upload operations, now it
is moved to PUT request with `action` parameter and have same format
like PUT /folders/{id}.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>