fix(copyright): do not update empty copyrights
Do not update the hash of copyrights where content is empty in 3.7-3.8
migration. Empty/NULL content and hash are used to denote no copyrights
found situations. The migration script updates the hash which causes
multiple collisions due to same hash.

Add new migration script dbmigrate_4.0-4.1.php to revert content and
hash to NULL if the hash was updated to `md5('')'

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>