For instance, if you want to upload files to the 'Fedora/ISOs/Disk 1' folder and add a description for the folder, you would use:
-f 'Fedora/ISOs/Disk 1' -d 'This folder contains the first disk of the Fedora distribution.'

Adding the -d option would set the folder description for the uploaded files. The tool could also provide an option to change the folder description for an existing folder in the FOSSology repository.

Overall, adding these options would enhance the usability of the cp2foss tool, making it more flexible and user-friendly.