feat(upload): track assignee and closing events
Implemented tracking of date-time on which a user was assigned to an
upload. Additionally, tracking has been extended to capture the
date-time when an upload was closed or rejected. This new feature will
enhance monitoring and auditing capabilities.

The values are currently available only on REST endpoints `GET /uploads`
and `GET /uploads/{id}`. The duration is displayed in Folder and upload
dashboard.

Signed-off-by: Gaurav Mishra <mishra.gaurav@siemens.com>