fix(fo_mapping_license.php): Fix update_license()
When upgrading from version < 4.3 to 4.3 or later, obsolete identifiers will be
renamed (e.g. GPL-2.0 -> GPL-2.0 only). If a license candidate exists with a
shortname identical to the one to which we want to rename, edited results and
the main license can be lost for existing uploads. This happens, because
update_license() is not taking all relevant tables into account (only the
license_file table is touched). Let's update clearing_event, license_set_bulk
and upload_clearing_license as well.

Signed-off-by: Jan Altenberg <jan.altenberg@osadl.org>