fix(Report): Check the return value of getLicenseById()
getCachedLicenseRisk() and getCachedLicenseText() do call getRisk() /
getText() on $this->licenseCache[$licenseId] without checking the value
before. Let's check $this->licenseCache[$licenseId] for null before
calling getRisk() or getText(). In case $this->licenseCache[$licenseId]
is null, we return null to the callers of getCachedLicenseRisk() and
getCachedLicenseText(), so they can deal with it.

Signed-off-by: Jan Altenberg <jan.altenberg@osadl.org>