fix(spdx2): Mark duplicates correctly with LicenseRef prefix
The intention of deduplicateLicenseList() is de-duplicating licenses
with the same SPDX ID. Duplicates are marked by adding a LicenseRef
prefix and adding the MD5 hash of the license text to the shortname. The
current sanity check before adding the LicenseRef prefix is not
correct: The sanity check has to call getShortName() instead of
getSpdxId().

Signed-off-by: Jan Altenberg <jan.altenberg@osadl.org>