fix(reportImport): (Re-)enable import of SPDX 2.2 documents
Since merging the support for SPDX version 3.0 it's not possible anymore
to import SPDX version 2.2 based documents. The reason is an explicit
check for version 2.3. SPDX 2.2 could be imported so far and no changes
to the relevant code have been done with version 3.0 support. So, let's
allow imports of SPDX 2.2 based documents again.

Signed-off-by: Jan Altenberg <jan.altenberg@osadl.org>