fix(spdxutils): improve handling of dual-license for spdxtv
implodeLicenses() explicitely parses for "Dual-license". If a user decides to
add a custom text for the Dual-license conclusion the license name will be
Dual-license-MD5. In this case dual licensing is not reflected correctly in the
SPDX2TV report. Change the behaviour of implodeLicenses to look for the prefix
Dual-license to fix the generation of SPDX2TV reports for Dual-license with
custom text.

Signed-off-by: Jan Altenberg <jan.altenberg@osadl.org>