fix(spdxutils): fix export of spdxtv reports
implodeLicenses() has been changed to be able to deal with Dual-license and
custom text. This commit introduced a bug, which breaks the export of SPDXTV
reports (handling of AND statements is not done correctly). This change fixes the
issue, restoring the original behaviour and keeping the ability of handling
Dual-license with custom text.

Signed-off-by: Jan Altenberg <jan.altenberg@osadl.org>